#109 renamed all instances of puterbot to openadapt
