+++ alembic/env.py
@@ -18,9 +18,9 @@ if config.config_file_name is not None:
 # for 'autogenerate' support
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
-from puterbot.config import DB_URL
-from puterbot.models import *
-from puterbot.db import Base
+from openadapt.config import DB_URL
+from openadapt.models import *
+from openadapt.db import Base
 target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
