+++ openadapt/common.py
