+++ openadapt/config.py
@@ -1,7 +1,7 @@
 import pathlib
 
 ROOT_DIRPATH = pathlib.Path(__file__).parent.parent.resolve()
-DB_FNAME = "puterbot.db"
+DB_FNAME = "openadapt.db"
 
 DB_FPATH = ROOT_DIRPATH / DB_FNAME
 DB_URL = f"sqlite:///{DB_FPATH}"
