+++ openadapt/db.py
@@ -4,8 +4,8 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.schema import MetaData
 from sqlalchemy.ext.declarative import declarative_base
 
-from puterbot.config import DB_ECHO, DB_URL
-from puterbot.utils import EMPTY, row2dict
+from openadapt.config import DB_ECHO, DB_URL
+from openadapt.utils import EMPTY, row2dict
 
 
 NAMING_CONVENTION = {
