+++ openadapt/playback.py
@@ -3,7 +3,7 @@
 from loguru import logger
 from pynput import mouse
 
-from puterbot.common import KEY_EVENTS, MOUSE_EVENTS
+from openadapt.common import KEY_EVENTS, MOUSE_EVENTS
 
 
 def play_mouse_event(event, mouse_controller):
