+++ openadapt/replay.py
@@ -1,8 +1,8 @@
 from loguru import logger
 import fire
 
-from puterbot.crud import get_latest_recording
-from puterbot.utils import configure_logging, get_strategy_class_by_name
+from openadapt.crud import get_latest_recording
+from openadapt.utils import configure_logging, get_strategy_class_by_name
 
 
 LOG_LEVEL = "INFO"
