+++ openadapt/visualize.py
@@ -9,13 +9,13 @@ from bokeh.layouts import layout, row
 from bokeh.models.widgets import Div
 from loguru import logger
 
-from puterbot.crud import (
+from openadapt.crud import (
     get_latest_recording,
 )
-from puterbot.events import (
+from openadapt.events import (
     get_events,
 )
-from puterbot.utils import (
+from openadapt.utils import (
     configure_logging,
     display_event,
     evenly_spaced,
