--- puterbot/strategies/__init__.py
@@ -1,4 +1,0 @@
-from puterbot.strategies.base import BaseReplayStrategy
-from puterbot.strategies.naive import NaiveReplayStrategy
-from puterbot.strategies.demo import DemoReplayStrategy
-# add more strategies here
